Use waitid in WaitContext
WaitContext doesn't reap the process (as before) thanks to WNOWAIT flag.
In the future, it might make sense to customize this behaviour.

Closes #1
