+++ pidfd_others.go
@@ -3,6 +3,7 @@
 package pidfd
 
 import (
+	"context"
 	"errors"
 	"fmt"
 	"os"
@@ -20,6 +21,7 @@ func open(_ int) (*File, error) { return nil, errUnimplemented }
 type conn struct{}
 
 func (*File) sendSignal(_ os.Signal) error { return errUnimplemented }
+func (*File) wait(_ context.Context) error { return errUnimplemented }
 
 func (*conn) Close() error                      { return errUnimplemented }
 func (*conn) SetReadDeadline(_ time.Time) error { return errUnimplemented }
